New shadow way of execution
It uses functions that Shadow has yet to implement and is generally not
finished yet but it contains enough functionality to run on its own outside of
Shadow, neatly emulating the Python interpreter. The code footprint could be
greatly reduced under Shadows new execution model, re-using Pythons own entry
functions.

However, on the flipside, we now have our own interposer and require either
Shadow to load it for us as part of the plugin or need to explicitly specify it
when running a simulation.